Implemented format-agnostic lowering algorithm

+++ include/taco/lower/lower.h
@@ -22,7 +22,7 @@ enum Property {
 /// Lower the tensor object with a defined expression and an iteration schedule
 /// into a statement that evaluates it.
 ir::Stmt lower(TensorVar tensor, std::string functionName,
-               std::set<Property> properties, int allocSize);
+               std::set<Property> properties, long long allocSize);
 
 }}
 #endif
